refactor: rename file to fit functions it exports
