+++ src/libs/algolia/fetchAlgolia/fetchMultipleVenues.ts
@@ -7,7 +7,7 @@ import { VenuesFacets } from 'libs/algolia/enums'
 import { isVenueHitTypeguard } from 'libs/algolia/fetchAlgolia'
 import { captureAlgoliaError } from 'libs/algolia/fetchAlgolia/AlgoliaError'
 import { client } from 'libs/algolia/fetchAlgolia/clients'
-import { buildGeolocationParameter } from 'libs/algolia/fetchAlgolia/fetchAlgolia'
+import { buildGeolocationParameter } from 'libs/algolia/fetchAlgolia/fetchOffer'
 import { buildHitsPerPage } from 'libs/algolia/fetchAlgolia/utils'
 import { env } from 'libs/environment'
 import { GeoCoordinates } from 'libs/geolocation'
